CORE-12218: Change `JsonObjectSerializer` to handle JSON
It fixes previously introduced test.
Also, some refactoring of JSON serialization/deserialization code.
