+++ libs/rest/rest-common/src/main/java/net/corda/rest/json/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.rest.json;
+
+import org.osgi.annotation.bundle.Export;
