+++ libs/rest/rest-common/src/main/kotlin/net/corda/rest/json/impl/JsonObjectAsString.kt
@@ -1,11 +1,11 @@
-package net.corda.rest.server.impl.internal
+package net.corda.rest.json.impl
 
 import net.corda.rest.JsonObject
 
 /**
  * Implementation of [JsonObject] that provides the [escapedJson] of a Json object for marshalling purposes.
  */
-data class JsonObjectAsString(override val escapedJson: String) : JsonObject {
+internal data class JsonObjectAsString(override val escapedJson: String) : JsonObject {
     override fun toString(): String {
         return escapedJson
     }
