+++ libs/rest/rest/src/main/kotlin/net/corda/rest/JsonObject.kt
@@ -1,8 +1,8 @@
 package net.corda.rest
 
 /**
- * Interface which can be used in HTTP request / response objects to indicate a field is a JSON value, object or array. Request payloads
- * will unmarshall to an escaped string which can be obtained through [escapedJson].
+ * Interface which can be used in HTTP request / response objects to indicate a possibility of a field being a JSON value or a string.
+ * Request payloads will unmarshall it to an escaped string which can be obtained through [escapedJson].
  *
  * This allows real JSON to be supplied in requests rather than escaping JSON and supplying it as a string value field.
  *
@@ -28,6 +28,9 @@ package net.corda.rest
  *   "obj": "{\"message\":\"Hey Mars\", \"planetaryOnly\":\"true\", \"target\":\"C=GB, L=FOURTH, O=MARS, OU=PLANET\"}"
  * }
  * ```
+ *
+ * In a similar way if this interface is used in the response and [escapedJson] represents a valid JSON - it will be
+ * included in the response verbatim, i.e. without escaping.
  */
 interface JsonObject {
     val escapedJson: String
